Move string difference functions to jest-diff package
