+++ .eslintignore
@@ -4,7 +4,7 @@ bin/
 flow-typed/**
 packages/*/build/**
 packages/*/build-es5/**
-packages/jest-matcher-utils/src/cleanupSemantic.ts
+packages/jest-diff/src/cleanupSemantic.ts
 website/blog
 website/build
 website/node_modules
