+++ packages/jest-diff/src/__tests__/getAlignedDiffs.test.ts
