+++ packages/jest-diff/src/joinAlignedDiffs.ts
