+++ scripts/checkCopyrightHeaders.js
@@ -101,7 +101,7 @@ const CUSTOM_IGNORED_PATTERNS = [
   '^flow-typed/.*',
   '^packages/expect/src/jasmineUtils\\.ts$',
   '^packages/jest-config/src/vendor/jsonlint\\.js$',
-  '^packages/jest-matcher-utils/src/cleanupSemantic\\.ts$',
+  '^packages/jest-diff/src/cleanupSemantic\\.ts$',
 ].map(createRegExp);
 
 const IGNORED_PATTERNS = [
